Renamed CourseSelector -> ExamSelector. Moved page titles from the pages int o components for reusability.
