--- src/features/courses/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./CourseCard";
-export * from "./CourseSelector";
