--- src/features/courses/store/coursesState.ts
@@ -1,20 +1,0 @@
-import { useDispatch, useSelector } from "react-redux";
-import { selectIsLoading, selectAllCourses } from "./coursesSelector";
-import { fetchCourses } from "./coursesSlice";
-
-export const useCoursesState = () => {
-  const dispatch = useDispatch();
-
-  const isLoading = useSelector(selectIsLoading);
-  const courses = useSelector(selectAllCourses);
-
-  const load = () => {
-    dispatch(fetchCourses());
-  };
-
-  return {
-    load,
-    isLoading,
-    courses
-  };
-};
