--- src/features/courses/store/index.ts
@@ -1,3 +1,0 @@
-export * from "./coursesSelector";
-export * from "./coursesSlice";
-export * from "./coursesState";
