+++ src/features/exams/components/index.ts
@@ -1,2 +1,4 @@
 export * from "./ExamConfigurator";
+export * from "./ExamSelector";
+export * from "./ExamSelectorCard";
 export * from "./ExamSimulator";
