+++ src/redux/rootReducer.ts
@@ -1,10 +1,8 @@
 import { combineReducers } from "@reduxjs/toolkit";
 
-import coursesReducer from "features/courses/store/coursesSlice";
 import examReducer from "features/exams/store/examSlice";
 
 const rootReducer = combineReducers({
-  courses: coursesReducer,
   exam: examReducer
 });
